Move kernel boot handling to new internal trait
