+++ packages/framework/src/Foundation/Concerns/BootsHydeKernel.php
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Foundation\Concerns;
 
+use Hyde\Foundation\FileCollection;
+use Hyde\Foundation\PageCollection;
+use Hyde\Foundation\RouteCollection;
+
 /**
  * @internal Single-use trait for the HydeKernel class.
  *
@@ -11,5 +15,12 @@ namespace Hyde\Foundation\Concerns;
  */
 trait BootsHydeKernel
 {
-    //
+    public function boot(): void
+    {
+        $this->booted = true;
+
+        $this->files = FileCollection::boot($this);
+        $this->pages = PageCollection::boot($this);
+        $this->routes = RouteCollection::boot($this);
+    }
 }
